separating out main grid layout into separate components and functions for readability
